+++ icekit/dashboard/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = '%s.apps.DashboardConfig' % __name__  # Portable