+++ icekit/plugins/image/admin.py
@@ -6,7 +6,7 @@ from django.utils.safestring import mark_safe
 from icekit.content_collections.admin import TitleSlugAdmin
 from icekit.plugins.image.models import ImageRepurposeConfig
 
-from icekit.utils.admin.mixins import ThumbnailAdminMixin
+from icekit.admin_tools.mixins import ThumbnailAdminMixin
 
 from . import models
 
@@ -16,8 +16,8 @@ except (AttributeError, KeyError):
     ADMIN_THUMB_ALIAS = {'size': (150, 150)}
 
 class ImageAdmin(ThumbnailAdminMixin, admin.ModelAdmin):
-    list_display = ['thumbnail', 'title', 'alt_text', 'is_ok_for_web', 'date_created', 'date_modified']
-    list_display_links = ['alt_text', 'title', 'thumbnail']
+    list_display = ['preview', 'title', 'alt_text', 'is_ok_for_web', 'date_created', 'date_modified']
+    list_display_links = ['alt_text', 'title', 'preview']
     date_hierarchy = 'date_modified'
     filter_horizontal = ['categories', ]
     list_filter = ['categories',]