+++ icekit/plugins/slideshow/admin.py
@@ -2,12 +2,15 @@ from django.contrib import admin
 from fluent_contents.admin import PlaceholderFieldAdmin
 
 from icekit.admin import ICEkitContentsAdmin
+from icekit.admin_tools.mixins import PreviewAdminMixin
 
 from . import models
 
 
-class SlideShowAdmin(PlaceholderFieldAdmin, ICEkitContentsAdmin):
+class SlideShowAdmin(PlaceholderFieldAdmin, ICEkitContentsAdmin, PreviewAdminMixin):
     change_form_template = 'icekit/plugins/slideshow/admin/change_form.html'
+    list_display = ICEkitContentsAdmin.list_display[:1] + ('preview', ) + ICEkitContentsAdmin.list_display[1:]
+    list_display_links = ('preview', ) + ICEkitContentsAdmin.list_display_links
     fieldsets = (
         (
             None, {